// src/dag/types/style-tokens.ts
export type NamedStyle =
  | "TITLE"
  | "SUBTITLE"
  | "HEADING_1"
  | "HEADING_2"
  | "HEADING_3"
  | "HEADING_4"
  | "HEADING_5"
  | "HEADING_6"
  | "NORMAL_TEXT";

export type Alignment = "START" | "CENTER" | "END" | "JUSTIFIED";

export type ListStyle = "BULLET" | "NUMBERED" | "CHECK";

export type ImageWrapLayout =
  | "WRAP_TEXT"
  | "BREAK_LEFT"
  | "BREAK_RIGHT"
  | "BREAK_LEFT_RIGHT"
  | "IN_FRONT_OF_TEXT"
  | "BEHIND_TEXT";